Basic Authentication class

This class handles the authentication to an openEO back-end that supports "basic" as login type. The class handles the retrieval of an access token by sending the encoded token consisting of user name and the password via HTTP header 'Authorization'. The authentication will be done once via login or multiple times when the lease time runs out. This class is created and registered in the OpenEOClient. After the login the user_id and the access_token are obtained and used as "bearer token" for the password restricted webservices.

Value

an object of type R6Class representing basic authentication

Details

The class inherits all fields and function from IAuth

Methods

$new(endpoint,user,password)

the constructor with the login endpoint and the credentials

Arguments

endpoint

the basic authentication endpoint as absolute URL

user

the user name

password

the user password