Graph object

This class represents an openEO process graph - which is generally denoted as field process_graph in the exchange objects of the API. The graph consists of ProcessNodes and optional ProcessGraphParameter (former variables). The explicit creation of a Graph is usually not required and discouraged, because this will be handled automatically.

Value

Object of R6Class with methods for building an openEO process graph

Details

In terms of the openEO API the process graph is the technical description of a process. To create a user-defined process it requires a process graph and additional meta data. The process graph is not accepted at any openEO endpoint directly. Therefore, it has to be wrapped in a Process object. Use as.Process in those cases. It is similarly handled in other functions of this package.

Fields

data

a named list of collection ids or process graph parameters depending on the context

Methods

$new(con = NULL, final_node=NULL)

The object creator created from processes and available data.

$getNodes()

a function to return a list of created ProcessNodes for this graph

$serialize()

creates a list representation of the graph by recursively calling $serialize

$validate()

runs through the nodes and checks the validity of its argument values

$getNode(node_id)

searches and returns a node from within the graph referenced by its node id

$addNode(node)

adds a ProcessNode to the graph

$removeNode(node_id)

removes a process node from the graph

$getFinalNode()

gets the result process node of a process graph

$setFinalNode(node)

sets the result process node by node id or a ProcessNode

$getVariables()

creates a named list of the defined variables of a process graph

$setVariables(list_of_vars)

sets the ProcessGraphParameter ( former variables) of graph

Arguments

con

openeo connection (optional) otherwise active_connection is used

final_node

optional, the final node (end node) that was used to create a graph

node_id

the id of a process node

node

process node or its node id

parameter

the name of a parameter in a process

value

the value to be set for a parameter of a paricular process

id or variable_id

the variable id

description

a description field for a variable

type

the type of variable, default 'string'

default

optional default value to be set for a variable