MetadataFilter

Inheriting from ProcessGraphArgument in order to represent a list of functions that is internally interpreted into Process() objects.

Value

Object of R6::R6Class() representing a list of Process() in order to filter for collections.

See also

Examples

if (FALSE) { # \dontrun{
# define filter statement
filter = list(
   "eo:cloud_cover" = function(x) x >= 0 & x < 50, 
   "platform" = function(x) x == "Sentinel-2A"
)

# setting the arguments is done via the process graph building with of 'processes()'
} # }