openEO client class

An R6Class that interacts with an openEO compliant back-end.

Fields

user_id

The user_id obtained after authentication

api.mapping

The mapping of the API endpoints and the back-end published ones

processes

a list of Process objects offered by the back-end

Methods

$new(host=NULL)

the constructor with an optional host URL to connect to

$getBackendEndpoint(endpoint_name)

returns the URL for the requested endpoint tag

$request(tag,parameters=NULL,authorized=FALSE, ...)

performs the desired HTTP request by endpoint tag with path parameters and whether or not authorization (access_token) is necessary

$isConnected()

whether or not the client has a host set

$isLoggedIn()

returns a logical describing whether the user is logged in

$getHost()

returns the host URL

$stopIfNotConnected()

throws an error if called and the client is not connected

$connect(url,version)

connects to a specific version of a back-end

$api_version()

returns the openEO API version this client is compliant to

$login(login_type = NULL,user=NULL,password=NULL)

creates an IAuth object based on the login_type

$logout()

invalidates the access_token and terminates the current session

$getAuthClient()

returns the authentication client

$setAuthClient(value)

sets the authentication client if it was configured and set externally

$getCapabilities()

service exploration to retrieve the supported openEO endpoints

$getDataCollection()

returns the list of collections as obtainable at 'list_collections()'

$getProcessCollection()

returns the evaluated process list as obtainable at 'processes()'

$getId()

returns the ID of the Connection as stated in the getCapabilities document

$getTitle()

returns the title of the connection as stated in the getCapabilities document

Arguments

host

the openEO host URL

endpoint_name

the endpoint tag the client uses for the endpoints

tag

endpoint tag

parameters

named list of values to be replaced in the endpoint

authorized

whether or not the endpoint requires authentication via access_token

url

url of an openEO back-end either directly versioned or with the separate version statement

version

the openEO API version to be used, or a list of available API versions if set to NULL

user

the user name

password

the user password

login_type

'basic', 'oidc' or NULL to control the authentication

value

an authentication object