ProcessGraphArgument

Inheriting from Argument in order to represent a ProcessGraph (prior known as callback). The ProcessGraph operates on reduced data of a data cube. For example reducing the time dimension results in a time series that has to be reduced into a single value. The value of a ProcessGraph is usually a Graph with ProcessGraphParameter as added data. Additional information can be found in the openEO API documention:

https://api.openeo.org/#section/Processes/Process-Graphs

Value

Object of R6Class representing a ProcessGraph.

Methods

$getProcessGraphParameters()

returns the available list ProcessGraphParameter

$setProcessGraphParameters(parameters)

assigns a list of ProcessGraphParameter to the ProcessGraph

Arguments

parameters

the ProcessGraphParameter list

See also