Connect to a openeEO service

Connects to openEO service. If the back-end provides a well-known endpoint that allows redirecting to specific versions you should provide the version parameter.

connect(
  host,
  version = NULL,
  user = NULL,
  password = NULL,
  login_type = NULL,
  exchange_token = "access_token",
  provider = NULL,
  config = NULL
)

Arguments

host
URL pointing to the openEO server service host

version
the openEO API version number as string (optional), see also api_versions

user
the user name (optional)

password
the password (optional)

login_type
either NULL, 'basic' or 'oidc'. This refers to the login mechanism that shall be used. NULL disables authentication.

exchange_token
'access_token' or 'id_token' defines in the OIDC case the bearer token use

provider
provider object as obtained by 'list_oidc_providers()'

config
named list containing 'client_id' and 'secret' or a path to the configuration file (type JSON)

Details

You can explore several already available openEO web services by using the openEO hub (https://hub.openeo.org/). There you have an overview about their status and connection details like the URL and supported features. You can explore the service for free through the access to publicly available metadata of data collections as well as the offered processing functions. For any computation and the creation of webservices, you need to register the openEO partner of your choice. There you will get further information on credentials and the login procedure.

Especially the login_type and the authType suggested by the client development guidelines are confusing. Here the login_type deals with considered login. 'Basic' allows you to use username and password directly in the call, whereas 'oidc' will open a browser window, where you enter you credentials. The authentication on all protected endpoints will later use the bearer token that the client has obtained after the login, unless the authentication was dropped with NULL.

The parameter version is not required. If the service offers a well-known document of the service the client will choose an appropriate version (default the most recent production ready version).

When calling this function the OpenEOClient is also stored in a variable in the package which marks the latest service that was connected to.

See also

Examples

if (FALSE) {
# The following examples show different configuration settings and point 
# to imaginary URLs. Please obtain a valid URL via the openEO hub and 
# register with one of the provider if required.

# connect to a host of the latest version and without authentication
con = connect(host='http://example.openeo.org')

# connect to a host by direct url and basic login
con = connect(host='http://example.openeo.org/v1.0',
              user='user',
             password='password',
             login_type='basic')

# connect to a host with open id connect authentication
con = connect(host='http://example.openeo.org',
              login_type='oidc')
}