Creates a new job on the back-end

In preparation to execute the users analysis workflow (user defined process) asynchronously, they need to register a job that will be scheduled when the required resources are available. To do so the user provides the process graph with optional descriptive meta data and the desired execution plan or the maximum amount of credits spent.

create_job(
  graph = NULL,
  title = NULL,
  description = NULL,
  plan = NULL,
  budget = NULL,
  con = NULL,
  ...
)

Arguments

graph

A Graph(), a function returning a ProcessNode() as an endpoint or the ProcessNode() will return the results

title

Optional title of a job

description

Optional detailed information about a job

plan

An optional execution plan offered by the back-end, determining how the job will be executed

budget

An optional budget, which sets the maximum amount of credits to be used by the job

con

connected and authenticated openEO client (optional) otherwise active_connection() is used.

...

additional parameters passed to jsonlite::toJSON() (like 'digits')

Value

the id of the job