Group operator wrappers

R's mathematical group primitives that are mapped into openEO processes.

# S3 method for ProcessNode
sum(..., na.rm = FALSE)

# S3 method for ProcessGraphParameter
sum(..., na.rm = FALSE)

# S3 method for list
sum(..., na.rm = FALSE)

# S3 method for ProcessNode
prod(..., na.rm = TRUE)

# S3 method for ProcessGraphParameter
prod(..., na.rm = TRUE)

# S3 method for list
prod(..., na.rm = TRUE)

# S3 method for ProcessNode
min(..., na.rm = TRUE)

# S3 method for ProcessGraphParameter
min(..., na.rm = TRUE)

# S3 method for list
min(..., na.rm = TRUE)

# S3 method for ProcessNode
max(..., na.rm = TRUE)

# S3 method for ProcessGraphParameter
max(..., na.rm = TRUE)

# S3 method for list
max(..., na.rm = TRUE)

# S3 method for ProcessNode
range(..., na.rm = TRUE)

# S3 method for ProcessGraphParameter
range(..., na.rm = TRUE)

# S3 method for list
range(..., na.rm = TRUE)

# S3 method for ProcessNode
mean(x, na.rm = FALSE, ...)

# S3 method for ProcessGraphParameter
mean(x, na.rm = FALSE, ...)

# S3 method for list
mean(x, na.rm = FALSE, ...)

# S3 method for ProcessNode
median(x, na.rm = FALSE, ...)

# S3 method for ProcessGraphParameter
median(x, na.rm = FALSE, ...)

# S3 method for list
median(x, na.rm = FALSE, ...)

sd.ProcessNode(x, na.rm = FALSE)

sd.ProcessGraphParameter(x, na.rm = FALSE)

sd.list(x, na.rm = FALSE)

var.ProcessNode(x, na.rm = FALSE)

var.ProcessGraphParameter(x, na.rm = FALSE)

var.list(x, na.rm = FALSE)

Arguments

...
multiple arguments that start with a ProcessNode or a ProcessGraphParameter

na.rm
logical to determine if NA values shall be removed in the calculation or if they are propagated

x
a vector or list of values that are mixed or consits fully of ProcessNode, ProcessGraphParameter or numerical values

Value

ProcessNode