List data on connected server

List available collections stored on an openEO server and return them as a CollectionList - a named list of Collection objects. The names are the collection IDs. Although the result at describe_collection() is also a Collection, the Collection object of returned from list_collections() is considered a list entry with less detailed information. The collection list is stored internally as the package environment variable data_collection, which can be accessed and set with active_data_collection().

list_collections(con = NULL)

active_data_collection(collection = NULL)

Arguments

con

Connection object (optional) otherwise active_connection() is used.

collection

the 'CollectionList' object of list_collections to be set as the active data collection in the package environment or left empty or NULL to return the package environment variable.

Value

object of class 'CollectionList'