Visualize the terms of service

If the service provides information about their terms of service in the capabilities, the function opens a new RStudio viewer panel and visualizes the HTML content of the link.

terms_of_service(con = NULL)

Arguments

con

a connected openEO client object (optional) otherwise active_connection() is used.

Value

a list of the link identifying the terms of service from the service capabilities or NULL