Unary function wrappers

The functions here are used in combination with ProcessGraphParameter and ProcessNode and facilitate writing arithmetic functions for openEO user defined processes in R. The functions translate into their openEO processes counterparts.

# S3 method for ProcessNode
abs(x)

# S3 method for ProcessGraphParameter
abs(x)

# S3 method for ProcessNode
sign(x)

# S3 method for ProcessGraphParameter
sign(x)

# S3 method for ProcessNode
sqrt(x)

# S3 method for ProcessGraphParameter
sqrt(x)

# S3 method for ProcessNode
trunc(x, ...)

# S3 method for ProcessGraphParameter
trunc(x, ...)

# S3 method for ProcessNode
floor(x)

# S3 method for ProcessGraphParameter
floor(x)

# S3 method for ProcessNode
ceiling(x)

# S3 method for ProcessGraphParameter
ceiling(x)

# S3 method for ProcessNode
round(x, digits = 0)

# S3 method for ProcessGraphParameter
round(x, digits = 0)

# S3 method for ProcessNode
exp(x)

# S3 method for ProcessGraphParameter
exp(x)

# S3 method for ProcessNode
log(x, base = exp(1))

# S3 method for ProcessGraphParameter
log(x, base = exp(1))

log10.ProcessNode(x)

log10.ProcessGraphParameter(x)

# S3 method for ProcessNode
cos(x)

# S3 method for ProcessGraphParameter
cos(x)

# S3 method for ProcessNode
sin(x)

# S3 method for ProcessGraphParameter
sin(x)

# S3 method for ProcessNode
tan(x)

# S3 method for ProcessGraphParameter
tan(x)

# S3 method for ProcessNode
cosh(x)

# S3 method for ProcessGraphParameter
cosh(x)

# S3 method for ProcessNode
sinh(x)

# S3 method for ProcessGraphParameter
sinh(x)

# S3 method for ProcessNode
tanh(x)

# S3 method for ProcessGraphParameter
tanh(x)

# S3 method for ProcessNode
acos(x)

# S3 method for ProcessGraphParameter
acos(x)

# S3 method for ProcessNode
asin(x)

# S3 method for ProcessGraphParameter
asin(x)

# S3 method for ProcessNode
atan(x)

# S3 method for ProcessGraphParameter
atan(x)

# S3 method for ProcessNode
acosh(x)

# S3 method for ProcessGraphParameter
acosh(x)

# S3 method for ProcessNode
asinh(x)

# S3 method for ProcessGraphParameter
asinh(x)

# S3 method for ProcessNode
atanh(x)

# S3 method for ProcessGraphParameter
atanh(x)

# S3 method for ProcessNode
cumsum(x)

# S3 method for ProcessGraphParameter
cumsum(x)

# S3 method for ProcessNode
cummin(x)

# S3 method for ProcessGraphParameter
cummin(x)

# S3 method for ProcessNode
cummax(x)

# S3 method for ProcessGraphParameter
cummax(x)

# S3 method for ProcessNode
cumprod(x)

# S3 method for ProcessGraphParameter
cumprod(x)

# S3 method for ProcessGraphParameter
[(x, i, ..., drop = TRUE)

# S3 method for ProcessNode
!(x)

# S3 method for ProcessGraphParameter
!(x)

# S3 method for ProcessNode
quantile(x, ...)

# S3 method for ProcessGraphParameter
quantile(x, ...)

Arguments

x
ProcessGraphParameter, ProcessNode or a list or vector. Passes internal data to the function

...
further arguments to pass on, see the documentation of primitive functions of R for further information

digits
the amount of decimal digits to round to

base
the base of the exponential operation

i
the index of the element in a vector or list

drop
listed for completeness but not used in openEO processes.

Value

a ProcessNode