Modifies a job with given parameter

The function modifies a stores a job with a given parameter. The dot parameter contains all the values that will be replaced or removed. The return shows a message of result or failure.

update_job(
  id,
  title = NULL,
  description = NULL,
  process = NULL,
  plan = NULL,
  budget = NULL,
  con = NULL,
  ...
)

Arguments

id

the job id of a created job

title

update title for the job

description

update description

process

A Graph(), a function returning a ProcessNode() as an endpoint, the ProcessNode() will return the results or a self defined Process()

plan

replaces plan with the set value

budget

replaces or sets the credits that can be spent at maximum

con

connected and authenticated openEO client (optional) otherwise active_connection() is used.

...

additional parameters passed to jsonlite::toJSON() (like 'digits')

Details

The '...' operator shall contain all the values that are to be replaced in the job. There are some reserved keys. The 'process_graph' option will replace the process graph with a newly defined one, therefore the process graph needs to be a Graph object. The 'format' option will change the desired output format. All other parameter will be assumed to be special output parameter. Remember, you don't need to specify a process graph or graph_id, e.g. if you just want to update the output format. To leave parameter unchanged, then don't mention it. If you want to delete some, then set them to NA.