OIDC Authentication

A class that authenticates via Open ID Connect. It inherits all fields and functions from IAuth. The OIDC login will open a browser window where you will be asked to enter your credentials. The website belongs to the OIDC provider of the chosen openeo back-end. Meanwhile the client will start a server demon in the background that listens for the callback from the OIDC provider. For this to work the user needs to get in contact with the openEO service provider and ask them for a configuration file that will contain information about the 'client_id' and 'secret'. The redirect URL requested from the provider is 'http://localhost:1410/'.

Details

The access_token will be returned when queried. If the lease time has run out the client will refresh the access_token automatically.

Methods

$new(provider, config=NULL)

the constructor for the authentication

$getUserData()

queries the OIDC provider for the user data like the 'user_id'

$getAuth()

returns the internal authentication client as created from package 'httr'

Arguments

provider

a provider object as returned by list_oidc_providers()

config

either a file to JSON containing information about 'client_id' and 'secret' or a named list